Explicitly only handle T and P from DG
Ignores other variables from the DG than T and P
